Make it react to user1/user2 button taps in the animations.

+++ PixelDriver/PHLevelsHorizAnimation.m
@@ -20,7 +20,7 @@
 
 - (void)renderBitmapInContext:(CGContextRef)cx size:(CGSize)size {
   if (self.driver.unifiedSpectrum) {
-    int wallQuartile = kWallHeight / 4;
+    CGFloat wallQuartile = kWallHeight / 4;
 
     CGContextSetRGBFillColor(cx, 1, 0, 0, 1);
     CGContextFillRect(cx, CGRectMake(0, 0, self.bassDegrader.value * size.width, wallQuartile));
